data/requirements.txt
jupyter
pandas
matplotlib
Pillow
idx2numpy
opencv-python